fetcher/requirements.txt
APScheduler==3.11.0
google-api-python-client==2.164.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
PyYAML==6.0.2
